+++ extensions/blocks/instagram-gallery/edit.js
@@ -30,7 +30,7 @@ import { addQueryArgs } from '@wordpress/url';
 import defaultAttributes from './attributes';
 import { getValidatedAttributes } from '../../shared/get-validated-attributes';
 import useConnectInstagram from './use-connect-instagram';
-import ImageFade from '../../shared/components/image-fade';
+import ImageTransition from './image-transition';
 import './editor.scss';
 
 const InstagramGalleryEdit = props => {
@@ -130,7 +130,7 @@ const InstagramGalleryEdit = props => {
 							key={ image.title || image.link }
 							style={ photoStyle }
 						>
-							<ImageFade src={ image.url } opacityTransition="1" photoStyle={ photoStyle } />
+							<ImageTransition src={ image.url } attributes={ attributes } />
 						</span>
 					) ) }
 					{ isLoadingGallery && count > images.length && (
