+++ extensions/blocks/instagram-gallery/editor.scss
@@ -9,6 +9,7 @@
 		box-sizing: border-box;
 		display: block;
 		line-height: 0;
+		position: relative;
 	}
 	img {
 		height: auto;
